Move assignment acceptance to offers
